Reserve sysid API for 32 bits (#3082)
* core: widen the system ID API to 32 bits

MAVLink 2 is getting 32 bit system IDs. The implementation is not ready
yet, but the public API should take the wider type before v4 so that
adding support later does not break it again.

This widens the system ID in Mavsdk::Configuration, System,
MavlinkAddress and MavlinkPassthrough, as well as the C API and the
Python and JNI bindings on top of it.

Internally, system IDs stay 8 bits. System IDs above 255 are rejected
with an error: a configuration falls back to the default system ID of
its component type, and passthrough commands to such a target fail.

* mavsdk_server: widen the system ID API to 32 bits

mavsdk_server_run_with_mavlink_ids() is used by the language wrappers,
so widen its system ID now as well, to avoid another API break once 32
bit system IDs are implemented.

The --sysid argument of the binary still only accepts 1..255.

* mavlink_passthrough: abort on out-of-range target_sysid in make_command_ack_message

target_sysid is now uint32_t, but make_command_ack_message() returns a
mavlink_message_t by value, so there's no Result to report a bad target
system ID through. Since sysid32 isn't supported yet, abort with a clear
error instead of silently truncating via static_cast<uint8_t>, matching the
existing abort_if_references_outlive_us() precedent for cases with no error
channel available.

* core: abort on out-of-range system ID in set_configuration and System::init

Both set_configuration_locked() and System::init() are void, so there is no
Result to report a bad system ID through. Previously they logged and
silently fell back (configuration) or no-op'd (System::init) instead of
aborting, which doesn't match the stated rule: log + return a Result where
possible, abort where there's no error channel. Aborts now, matching
make_command_ack_message's precedent.

* mavlink_direct: reject out-of-range target system/component IDs

MavlinkMessage.target_system_id/target_component_id are uint32 in the API,
but MAVLink system/component IDs are still 1 byte on the wire until sysid32
lands. Values over 255 were silently truncated by static_cast<uint8_t>
(e.g. 256 -> 0) instead of being rejected. Both send_message() paths
(mavlink_direct client and mavlink_direct_server) now return
Result::InvalidField for out-of-range values, with a system test covering
both paths.

* core: fix system ID doc comments in Mavsdk::Configuration

Out-of-range system IDs now abort when the configuration is applied
rather than falling back to the default system ID.

---------

+++ c/src/cmavsdk/system.h
@@ -26,7 +26,7 @@ CMAVSDK_EXPORT bool mavsdk_system_is_standalone(mavsdk_system_t system);
 CMAVSDK_EXPORT bool mavsdk_system_has_camera(mavsdk_system_t system, int camera_id);
 CMAVSDK_EXPORT bool mavsdk_system_has_gimbal(mavsdk_system_t system);
 CMAVSDK_EXPORT bool mavsdk_system_is_connected(mavsdk_system_t system);
-CMAVSDK_EXPORT uint8_t mavsdk_system_get_system_id(mavsdk_system_t system);
+CMAVSDK_EXPORT uint32_t mavsdk_system_get_system_id(mavsdk_system_t system);
 
 /**
  * @brief Get all component IDs.
